+++ packages/contracts/test/plugins/governance/multisig/multisig.ts
@@ -12,6 +12,7 @@ import {
   IMultisig__factory,
   IPlugin__factory,
   IProposal__factory,
+  IProtocolVersion__factory,
   Multisig,
   Multisig__factory,
 } from '../../../../typechain';
@@ -254,11 +255,11 @@ describe('Multisig', function () {
         currentContractFactory.attach(toImplementation)
       );
 
-      expect(fromProtocolVersion).to.deep.equal(toProtocolVersion); // The contracts inherited from OSx did not change from 1.0.0 to the current version
+      expect(fromProtocolVersion).to.not.deep.equal(toProtocolVersion);
       expect(fromProtocolVersion).to.deep.equal(
         IMPLICIT_INITIAL_PROTOCOL_VERSION
       );
-      expect(toProtocolVersion).to.not.deep.equal(CURRENT_PROTOCOL_VERSION);
+      expect(toProtocolVersion).to.deep.equal(CURRENT_PROTOCOL_VERSION);
     });
 
     it('from v1.3.0', async () => {
@@ -275,7 +276,7 @@ describe('Multisig', function () {
           UPGRADE_PERMISSIONS.UPGRADE_PLUGIN_PERMISSION_ID,
           dao
         );
-      expect(toImplementation).to.equal(fromImplementation);
+      expect(toImplementation).to.not.equal(fromImplementation);
 
       const fromProtocolVersion = await getProtocolVersion(
         legacyContractFactory.attach(fromImplementation)
@@ -284,15 +285,15 @@ describe('Multisig', function () {
         currentContractFactory.attach(toImplementation)
       );
 
-      expect(fromProtocolVersion).to.deep.equal(toProtocolVersion); // The contracts inherited from OSx did not change from 1.3.0 to the current version
+      expect(fromProtocolVersion).to.not.deep.equal(toProtocolVersion);
       expect(fromProtocolVersion).to.deep.equal(
         IMPLICIT_INITIAL_PROTOCOL_VERSION
       );
-      expect(toProtocolVersion).to.not.deep.equal(CURRENT_PROTOCOL_VERSION);
+      expect(toProtocolVersion).to.deep.equal(CURRENT_PROTOCOL_VERSION);
     });
   });
 
-  describe('plugin interface: ', async () => {
+  describe('ERC-165', async () => {
     it('does not support the empty interface', async () => {
       expect(await multisig.supportsInterface('0xffffffff')).to.be.false;
     });
@@ -305,6 +306,12 @@ describe('Multisig', function () {
 
     it('supports the `IPlugin` interface', async () => {
       const iface = IPlugin__factory.createInterface();
+      expect(await multisig.supportsInterface(getInterfaceID(iface))).to.be
+        .true;
+    });
+
+    it('supports the `IProtocolVersion` interface', async () => {
+      const iface = IProtocolVersion__factory.createInterface();
       expect(await multisig.supportsInterface(getInterfaceID(iface))).to.be
         .true;
     });