+++ packages/contracts/test/token/erc20/governance-wrapped-erc20.ts
@@ -11,6 +11,8 @@ import {
 } from '../../../typechain';
 import {OZ_ERRORS} from '../../test-utils/error';
 import {getInterfaceID} from '../../test-utils/interfaces';
+import {CURRENT_PROTOCOL_VERSION} from '../../test-utils/protocol-version';
+import {IProtocolVersion__factory} from '../../../typechain/@aragon/osx-v1.3.0-rc0.2/utils/protocol/IProtocolVersion.sol';
 
 export type AccountBalance = {account: string; amount: number};
 
@@ -109,13 +111,26 @@ describe('GovernanceWrappedERC20', function () {
     });
   });
 
-  describe('supportsInterface:', async () => {
+  describe('Protocol version', async () => {
+    it('returns the current protocol version', async () => {
+      expect(await governanceToken.protocolVersion()).to.deep.equal(
+        CURRENT_PROTOCOL_VERSION
+      );
+    });
+  });
+
+  describe('ERC-165', async () => {
     it('does not support the empty interface', async () => {
       expect(await governanceToken.supportsInterface('0xffffffff')).to.be.false;
     });
 
     it('supports the `IERC165Upgradeable` interface', async () => {
       const iface = IERC165Upgradeable__factory.createInterface();
+      expect(await governanceToken.supportsInterface(getInterfaceID(iface))).to
+        .be.true;
+    });
+    it('supports the `IProtocolVersion` interface', async () => {
+      const iface = IProtocolVersion__factory.createInterface();
       expect(await governanceToken.supportsInterface(getInterfaceID(iface))).to
         .be.true;
     });